refactor: useBookGroupQuery default export
