+++ src/ui/Group/GroupDetail/index.tsx
@@ -5,7 +5,7 @@ import GroupInfo from '@/ui/Group/GroupDetail/GroupInfo';
 import CommentInputBox from '../GroupComment/CommentInputBox';
 import CommentsList from '../GroupComment';
 import useGroupInfoQuery from '@/queries/group/useGroupInfoQuery';
-import useGroupCommentsQuery from '@/queries/group/useGroupCommentsQuery';
+import useGroupCommentsQuery from '@/queries/group/useBookGroupCommentsQuery';
 import GroupAPI from '@/apis/group';
 import { useToast } from '@/hooks/toast';
